Remove the spell check guard
The guard suspended checking when too much of the document came back
misspelled, to protect against a checker loaded with the wrong language. The
squiggle and relayout paths are fast enough now that flagging every word is no
longer the emergency it was written for, and the machinery it needed (sticky
suspension state, recovery paths through four entry points, a re-check on
every partial pass) cost more complexity than the case was worth.

Gone: SpellCheckGuard, SpellCheckSuspension, SpellCheckState.suspension,
resumeSpellChecking, and the guard parameters on SpellCheckState and
rememberSpellCheckState. None of it shipped in a release, so no published API
changes.

Kept from the same span of work, since none of it is guard machinery: the
batched updateRichSpans calls, the adjoining-range merge that keeps a typed
word to one check, checkWordSegment only decorating while checking is on, the
post-lookup re-checks that stop a swap undoing a concurrent disable, and the
e2e harness, now covering ordinary checking rather than the guard.
