+++ ComposeTextEditorSpellCheck/src/commonMain/kotlin/com/darkrockstudios/texteditor/spellcheck/rememberSpellCheckState.kt
@@ -13,15 +13,8 @@ import com.darkrockstudios.texteditor.state.rememberTextEditorState
  * Pass a STABLE [spellChecker] instance: `remember` it across recompositions and
  * only create a new one when the underlying dictionary actually changes. A fresh
  * instance every recomposition re-keys the full-rescan effect below and re-scans
- * the whole document on each frame. Re-scans are cancellation-safe, bail at the
- * guard's trip point, and never lift a suspension unless they complete with
- * plausible results, but the churn is pure waste.
- *
- * @param guard Sanity limits that suspend checking when its results stop looking plausible;
- *   see [SpellCheckGuard] and [SpellCheckState.suspension]. Changing [guard] or [spellChecker]
- *   on recomposition triggers a re-check that lifts any standing suspension when its results
- *   come back clean, since swapping the checker (or loosening the limits) is how a
- *   wrong-language suspension gets fixed.
+ * the whole document on each frame. Re-scans are cancellation-safe and won't lose
+ * spans, but the churn is pure waste.
  */
 @Composable
 fun rememberSpellCheckState(
@@ -29,32 +22,17 @@ fun rememberSpellCheckState(
 	initialText: AnnotatedString? = null,
 	enableSpellChecking: Boolean = true,
 	spellCheckMode: SpellCheckMode = SpellCheckMode.Word,
-	guard: SpellCheckGuard = SpellCheckGuard.Default,
 ): SpellCheckState {
 	val richTextState = rememberTextEditorState(initialText)
 	val state = remember {
-		SpellCheckState(richTextState, spellChecker, enableSpellChecking, spellCheckMode, guard)
+		SpellCheckState(richTextState, spellChecker, enableSpellChecking, spellCheckMode)
 	}
 
-	// Run SpellCheck as soon as it is ready. A full check scans even while suspended
-	// and lifts the suspension only on plausible results, so a replacement checker
-	// gets a fresh chance (it may well be the right-language one the previous
-	// suspension was asking for) without the banner flickering when it isn't.
+	// Run SpellCheck as soon as it is ready
 	LaunchedEffect(spellChecker) {
 		if (spellChecker != null) {
 			state.spellChecker = spellChecker
 			state.runFullSpellCheck()
-		}
-	}
-
-	// Propagate guard changes to the remembered state. A laxer guard can make a
-	// standing suspension obsolete, so re-check under the new limits.
-	LaunchedEffect(guard) {
-		if (state.guard != guard) {
-			state.guard = guard
-			if (state.suspension != null) {
-				state.resumeSpellChecking()
-			}
 		}
 	}
